Better color management.
I fixed the problem of red displaying as white. I noticed that COLOR_PAIR(0) cannot be overriden which is what I was doing for red. I've since found a better way to initialize the color pairs and obviates the Colors enum. Also added calls to wattron/wattroff throughout to reduce code duplication on printing.