Allow user to change code length.
Added getopt support to allow changing the number of guesses per round
and the length of the secret code to guess.

I'm not super satisfied with the hard-coded digit positions. However,
it works for now until I have time to figure out a way to position them
relative to the size of the containing window.