Trying out bandit.
After finding liblittletest wasn't going to work for my table-driven
tests I am going to try something newer. It seems like BDD-style test
frameworks are hot so lets give this a try. I am calling it quits for
the night without getting to actually implementing the full table test
(assuming it is feasible to do so with this framework.)